use parseFloat instead of parseInt to allow small donation amounts

+++ projects/bp-gallery/src/components/common/DonateButton.tsx
@@ -84,7 +84,7 @@ const DonateButton = ({
                 />
                 <Button
                   variant='outlined'
-                  onClick={() => (parseInt(donation) > 0 ? setIsOpen(!isOpen) : '')}
+                  onClick={() => (parseFloat(donation) > 0 ? setIsOpen(!isOpen) : '')}
                 >
                   {isOpen
                     ? t('archives.edit.paypal.donation-edit')
